ErrorTypeMatcher renamed to ErrorMatcher
