--- internal/pipeline/errorhandlers/matcher/error_type_matcher.go
@@ -1,17 +1,0 @@
-package matcher
-
-import (
-	"errors"
-)
-
-type ErrorTypeMatcher []error
-
-func (etm ErrorTypeMatcher) Match(err error) bool {
-	for _, v := range etm {
-		if errors.Is(err, v) {
-			return true
-		}
-	}
-
-	return false
-}
